改了点类型，删了macnet用不到的函数

+++ 2508date/src/MAC.hpp
@@ -51,7 +51,7 @@
 
 using namespace std;
 
-extern int packet_id;
+extern long long packet_id;
 
 extern unsigned int cycles;
